`NativePreviewView` -> `PreviewView`
